Fix mem leak issue
