+++ HybridCamera/VC.swift
@@ -6,6 +6,7 @@ import With
  * Main view controller
  */
 class VC: UIViewController {
+   var processMediaView: CustomProcessView?
    /**
     * We add the Camera view in the viewDidAppear so that its tricggered again after user changes the app prefs to allow video use
     */
@@ -24,46 +25,4 @@ class VC: UIViewController {
     */
    override var shouldAutorotate: Bool { return false }/*Locks autorotate*/
    override var preferredStatusBarStyle: UIStatusBarStyle { return .lightContent }
-}
-/**
- * Extension
- */
-extension VC {
-   /**
-    * When camera accesses is granted proced to initiate the camera
-    */
-   func initiate() {
-      let hybridCamView: CustomCamView = .init()
-      self.view = hybridCamView /*Add HybridCamView as the main view*/
-      hybridCamView.camView.onPhotoCaptureComplete = { [weak self] (image: UIImage?, url: URL?, error: Error?) in
-         guard let self = self else { Swift.print("mem leak"); return }
-         self.onCapture(image, url, error)
-      }
-      hybridCamView.camView.onVideoCaptureComplete = { [weak self] (url: URL?, error: Error?) in
-         guard let self = self else { Swift.print("mem leak"); return }
-         self.onCapture(nil, url, error)
-      }
-   }
-   /**
-    * When camera onCapture is called
-    */
-//   guard self = self else { return }
-   private func onCapture(_ image: UIImage?, _ url: URL?, _ error: Error?) {
-      weak var _self = self // temp fix for possible mem leak
-      guard let self = _self else { Swift.print("⚠️️ possible retian cycle ⚠️️"); return }
-      let processMediaView: CustomProcessView = {
-         let processMediaView: CustomProcessView = .init(frame: UIScreen.main.bounds)
-         processMediaView.onExit = {
-             processMediaView.deInitiate()
-         }
-         processMediaView.onShare = { (url: URL?) in if let url = url { CustomProcessView.promptSaveFileDialog(vc: self, url: url) { processMediaView.deInitiate() } } }
-         self.view.addSubview(processMediaView)
-         return processMediaView
-      }()
-      if let error = error {
-         CustomProcessView.promptErrorDialog(vc: self, error: error ) { processMediaView.deInitiate() }; return
-      } else {
-         processMediaView.present(image: image, url: url)
-      }
-   }
 }
